+++ CHANGES.md
@@ -13,6 +13,8 @@
 
 - Graphics:
   - New `Vertex Array` resource
+- Tests:
+  - New `Vertex Array` tests
 - Geometry:
   - New namespace `Coli::Geometry` for geometry features
   - New `Vertex` struct and its traits `VertexTraits`